pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yi-min-ai"
version = "0.1.0"
description = "Yi Min AI Assistant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "anthropic>=0.51.0",
  "openai>=1.30.0",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
  "pytest>=8.3.5",
  "pytest-asyncio>=0.26.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["agent"]

[tool.hatch.build.targets.wheel.force-include]
"config" = "config"
"workspace" = "workspace"
